% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integration}
\alias{integration}
\alias{enable}
\alias{disable}
\alias{enable_repo}
\alias{disable_repo}
\title{Enable/Disable RStudio Package Manager}
\usage{
enable()

disable()

enable_repo()

disable_repo()
}
\description{
Functions to enable or disable the integration of \code{\link{install_sysreqs}}
into \code{\link{install.packages}}. When enabled, binary packages are
installed from RSPM if available, and system requirements are transparently
resolved and installed without root privileges.
}
\details{
To enable \pkg{rspm} system-wide by default, include the following:

\code{suppressMessages(rspm::enable())}

into the \code{Rprofile.site} file. To enable it just for a particular user,
move that line to the user's \code{~/.Rprofile} instead.
}
\examples{
\dontrun{
# install 'units' and all its dependencies from the system repos
rspm::enable()
install.packages("units")

# install packages again from CRAN
rspm::disable()
install.packages("errors")
}

}
