% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_init}
\alias{renv_init}
\title{Initialize an \pkg{renv} Project}
\usage{
renv_init(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{renv::init()}.}
}
\description{
Substitutes \code{renv::init()} to initialize a new \pkg{renv} project with
\pkg{rspm} enabled. This function sets the appropriate infrastructure to
activate the integration. Then, packages can be installed normally via
\code{install.packages} and \code{update.packages}.
}
\details{
Note that, if \code{renv::install} or \code{renv::update} are called
directly, then \code{rspm::install_sysreqs()} needs to be called manually.
}
