% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{manager}
\alias{manager}
\alias{install_sysreqs}
\alias{set_rpath}
\title{Manage System Requirements}
\usage{
install_sysreqs()

set_rpath()
}
\description{
Functions to install and configure system requirements.
}
\seealso{
\code{\link{integration}}
}
