% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{manager}
\alias{manager}
\alias{install_sysreqs}
\title{Manage System Requirements}
\usage{
install_sysreqs()
}
\description{
Detect, install and configure system requirements. This function is
automatically called when the package is enabled via \code{\link{enable}}.
It can also be called manually at any time to update the system requirements.
}
