% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{manager}
\alias{manager}
\alias{install_sysreqs}
\alias{missing_sysreqs}
\title{Manage System Requirements}
\usage{
install_sysreqs()

missing_sysreqs()
}
\value{
\code{install_sysreqs}: No return value, called for side effects.

\code{missing_sysreqs}: A list of missing libraries, for debugging.
}
\description{
Detect, install and configure system requirements. This function is
automatically called when the package is enabled via \code{\link{enable}}.
It can also be called manually at any time to update the system requirements.
}
\examples{
\dontrun{
# install 'units' without enabling the integration
install.packages("units")
# then trigger the installation of system requirements manually
rspm::install_sysreqs()
}

}
