% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspm.R
\docType{package}
\name{rspm-package}
\alias{rspm-package}
\title{\pkg{rspm}: RStudio Package Manager}
\description{
Enables binary package installations on Linux distributions.
Provides access to RStudio public repositories at
\url{https://packagemanager.rstudio.com}, and transparent management of
system requirements without administrative privileges. Currently supported
distributions are CentOS / RHEL 7-9, and several RHEL derivatives
(Rocky Linux, AlmaLinux, Oracle Linux, Amazon Linux 2),
openSUSE / SLES 15.3-4, and Ubuntu 18.04, 20.04 and 22.04.
}
\references{
\url{https://enchufa2.github.io/rspm/}
}
\seealso{
\code{\link{manager}}, \code{\link{integration}}
}
\author{
Iñaki Ucar
}
