% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspm.R
\docType{package}
\name{rspm-package}
\alias{rspm-package}
\title{\pkg{rspm}: RStudio Package Manager}
\description{
Enables binary package installations on Linux distributions.
Provides access to RStudio public repositories at
\url{https://packagemanager.posit.co}, and transparent management of
system requirements without administrative privileges. Currently supported
distributions are CentOS / RHEL 7-9, and several RHEL derivatives
(Rocky Linux, AlmaLinux, Oracle Linux, Amazon Linux 2),
openSUSE / SLES 15.4-5, Debian 11-12, and Ubuntu LTS 20-22.
}
\references{
\url{https://cran4linux.github.io/rspm/}
}
\seealso{
\code{\link{manager}}, \code{\link{integration}}
}
\author{
Iñaki Ucar
}
